requirements.txt
gradio>=4.44.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0,<2.10
fastapi>=0.100.0,<0.115.0
requests>=2.31.0
typing-extensions>=4.12.0
openai>=1.0.0
geocoder>=1.38.1 